requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.5.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
environs==9.5.0
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
gspread==5.10.0
gspread-asyncio==1.9.0
httplib2==0.22.0
idna==3.4
magic-filter==1.0.11
marshmallow==3.20.1
multidict==6.0.4
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.1
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
PyDrive==1.3.1
pyparsing==3.1.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing_extensions==4.7.1
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.17
yarl==1.9.2
